+++ Mcode/+dispstrlib/Displayable.m
@@ -35,32 +35,46 @@ classdef Displayable
   methods
     
     function disp(this)
-      %DISP Custom display
-      disp(dispstr(this));
+      % Custom display
+      if isscalar(this)
+        disp(dispstr_scalar(this));
+      else
+        strs = dispstrs(this);
+        dispstrlib.DispstrHelper.disparray(strs);
+      end
     end
     
     function out = dispstr(this)
-      %DISPSTR Custom display string
+      % Custom display string for this array as a whole.
       if isscalar(this)
-        strs = dispstrs(this);
-        out = strs{1};
-      elseif ismatrix(this)
-        strs = dispstrs(this);
-        out = dispstrlib.DispstrHelper.disparray(strs);
+        out = dispstr_scalar(this);
       else
-%        out = sprintf('%s %s', dispstrlib.internal.DispstrImpl.size2str(size(this)), class(this));
-        strs = dispstrs(this);
-        out = dispstrlib.DispstrHelper.disparray(strs);
+        % Default to doing an opaque display, because we don't know if the
+        % object is going to spam whatever context it's in.
+        out = sprintf('<%s %s>', size2str(size(this)), class(this));
       end
     end
     
     function out = dispstrs(this)
-      out = cell(size(this));
+      % Element-wise custom display strings
+      %
+      % Gets the custom display strings for each element of this array, as
+      % opposed to a custom display string 
+      % Returns a string array the same size size as this.
+      out = repmat(string(missing), size(this));
       for i = 1:numel(this)
-        out{i} = dispstr_scalar(subsref(this, ...
+        out(i) = dispstr_scalar(subsref(this, ...
           struct('type','()', 'subs',{{i}})));
       end
     end
+    
+  end
+  
+  % Now here are some overrides that let you pass Displayables as arguments to
+  % Matlab's own error(), warning(), and related functions, and they'll
+  % auto-convert into strings that the %s conversion specifier can handle.
+  
+  methods
     
     function error(varargin)
       args = convertDisplayablesToString(varargin);